Better test experience
- setup a structure using [Jest](https://jestjs.io) to help asserting the solution
- added linting and auto formatting
