+++ package.json
@@ -1,0 +1,31 @@
+{
+  "name": "dsa-playground",
+  "version": "1.0.0",
+  "description": "A collection of Data Structures & Algorithms challenges.",
+  "repository": "https://github.com/mrmonsters/dsa-playground",
+  "author": "Chin Tiong",
+  "license": "MIT",
+  "type": "module",
+  "scripts": {
+    "lint": "eslint --ignore-path .eslintignore",
+    "postinstall": "husky",
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js"
+  },
+  "devDependencies": {
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "husky": "^9.0.11",
+    "jest": "^29.7.0",
+    "lint-staged": "^15.2.2",
+    "prettier": "^3.2.5"
+  },
+  "jest": {
+    "transform": {}
+  },
+  "lint-staged": {
+    "*.+(js)": [
+      "eslint --ignore-path .eslintignore --fix --quiet"
+    ]
+  }
+}
